Prvni verze stahovani ultimy z andarie (fixni URL a hash v Settings)

+++ trunk/AndariaPatcher/src/cz/polous/andaria/Installer.java
@@ -125,10 +125,13 @@ class Installer extends PatcherQueue {
         PatchItem patchItem = getFirstItem();
 
         // progress will be counted by 65% for extracting files
-        // and  35% for installing patches.
+        // and  35% for installing patches expect for uoml file.
         // ProgressBars - for single file
         resetProgressBar(BARS.SINGLE, patchItem.getSize());
-        extractProgressPart = Math.round(patchItem.getSize() * 65 / 100.00);
+        if (patchItem.getName().equals(Settings.getInstance().getUomlPatchItemName()))
+            extractProgressPart = patchItem.getSize();
+        else
+            extractProgressPart = Math.round(patchItem.getSize() * 65 / 100.00);
 
         log.addDebug("Pracuju se souborem: ".concat(patchItem.getLocalFileName()));
         log.addDebug("...jeho velikost je: ".concat(Double.toString(patchItem.getSize())));
@@ -137,7 +140,7 @@ class Installer extends PatcherQueue {
         if (patchItem.isPacked()) {
             final String uopath = Settings.getInstance().getValue(Settings.VALUES.ULTIMA_ONINE_PATH);
 
-            setLabelText("Rozbaluju patch: " + patchItem.getFileName());
+            setLabelText("Rozbaluju soubor: " + patchItem.getFileName());
             J7zipBinding sevenZip = new J7zipBinding();
             try {
                 sevenZip.unpack(patchItem.getLocalFileName());
@@ -170,6 +173,19 @@ class Installer extends PatcherQueue {
             if (f.exists()) {
                 log.addDebug("Našel jsem start_g.bat.");
                 setLabelText("Instaluju patch (start_g.bat): " + patchItem.getFileName());
+                exec(Settings.getInstance().getOs().getBatchExecCommand(f));
+                f.delete();
+                setSingleProgressPercents(97);
+            }
+
+            // - if exist start_g.bat, execute it
+            f = new File(uopath + File.separator + "install.bat");
+            setSingleProgressPercents(87);
+            log.addDebug("Hledám:" + f.getAbsolutePath());
+            if (canceled()) return;
+            if (f.exists()) {
+                log.addDebug("Našel jsem install.bat.");
+                setLabelText("Instaluju patch (install.bat): " + patchItem.getFileName());
                 exec(Settings.getInstance().getOs().getBatchExecCommand(f));
                 f.delete();
                 setSingleProgressPercents(97);
